fix(auth): capture auth_provider param on any route, not just /app
Bugbot review on #394: the oauth-complete bridge's return target can be
any same-origin path (deep links, subroutes like
/app/settings/notifications), but the auth_provider capture only ran in
the /app page mount effect, so logins landing on subroutes never
persisted the last-used provider and left the param in the URL. Hoist
the capture into LastAuthProviderCapture mounted in the root layout so
it runs wherever the redirect lands.
